requirements.txt
dateparser>=1.1.0
discord.py @ git+https://github.com/Baviaan/discord.py@f99f8b5baded82d513bb4da8d06a92bdfddbc569
fuzzywuzzy>=0.18.0
psutil>=5.8.0
pytz>=2021.3
requests>=2.27.0
